Test: Cable attached to plug
