+++ scripts/demos/cables.py
@@ -43,12 +43,23 @@ from isaaclab.assets import RigidObject, RigidObjectCfg
 
 from isaaclab_contrib.cable import CableAttachmentCfg, CableObject, CableObjectCfg
 
-PLUG_USDA = "/home/mmichelis/Documents/IsaacLab-Origin/scripts/demos/plug_mesh_flange_only.usda"
+PLUG_USDA = "/home/mmichelis/Documents/IsaacLab-Origin/scripts/demos/plug_mesh001.usda"
+CABLE_USDA = "/home/mmichelis/Documents/IsaacLab-Origin/scripts/demos/cable001.usda"
 
 
 def z_axis_quat(angle_rad: float) -> tuple[float, float, float, float]:
     """Quaternion (x, y, z, w) for a rotation of ``angle_rad`` about +Z."""
     return (0.0, 0.0, math.sin(0.5 * angle_rad), math.cos(0.5 * angle_rad))
+
+
+def y_axis_quat(angle_rad: float) -> tuple[float, float, float, float]:
+    """Quaternion (x, y, z, w) for a rotation of ``angle_rad`` about +Y."""
+    return (0.0, math.sin(0.5 * angle_rad), 0.0, math.cos(0.5 * angle_rad))
+
+
+def x_axis_quat(angle_rad: float) -> tuple[float, float, float, float]:
+    """Quaternion (x, y, z, w) for a rotation of ``angle_rad`` about +X."""
+    return (math.sin(0.5 * angle_rad), 0.0, 0.0, math.cos(0.5 * angle_rad))
 
 
 def design_scene(num_cables: int) -> dict[str, "CableObject | RigidObject"]:
@@ -70,8 +81,10 @@ def design_scene(num_cables: int) -> dict[str, "CableObject | RigidObject"]:
     entities: dict[str, CableObject | RigidObject] = {}
     for idx in tqdm.tqdm(range(num_cables)):
         angle = random.uniform(0.0, 2.0 * math.pi)
-        cx = random.uniform(-xy_jitter, xy_jitter) - 0.5 * cable_length * math.cos(angle)
-        cy = random.uniform(-xy_jitter, xy_jitter) - 0.5 * cable_length * math.sin(angle)
+        # cx = random.uniform(-xy_jitter, xy_jitter) - 0.5 * cable_length * math.cos(angle)
+        # cy = random.uniform(-xy_jitter, xy_jitter) - 0.5 * cable_length * math.sin(angle)
+        cx = 0.0
+        cy = 0.0
         cz = z_base + idx * z_spacing
 
         # Tail-body world position (the last edge body sits one segment back from the tail node).
@@ -83,20 +96,16 @@ def design_scene(num_cables: int) -> dict[str, "CableObject | RigidObject"]:
             prim_path=plug_prim_path,
             spawn=sim_utils.UsdFileCfg(usd_path=PLUG_USDA),
             init_state=RigidObjectCfg.InitialStateCfg(
-                pos=(tail_body_x, tail_body_y, cz),
-                rot=z_axis_quat(angle),
+                pos=(-0.38398558, 0.34585292, 0.5-0.36874688),
+                rot=(0.0, -0.57096256, 0.0, 0.8209761),
             ),
         )
-        entities[f"Plug{idx:03d}"] = RigidObject(cfg=plug_cfg)
+        entities[f"Plug1{idx:03d}"] = RigidObject(cfg=plug_cfg)
 
         cable_cfg = CableObjectCfg(
-            prim_path=f"/World/Origin/Cable{idx:03d}",
-            spawn=sim_utils.CableCfg(
-                positions=[(i * segment_length, 0.0, 0.0) for i in range(num_points)],
-                width=width,
-                visual_material=sim_utils.PreviewSurfaceCfg(
-                    diffuse_color=(random.random(), random.random(), random.random())
-                ),
+            prim_path=f"/World/Origin/Cable1{idx:03d}",
+            spawn=sim_utils.UsdFileCfg(
+                usd_path=CABLE_USDA,
                 physics_material=NewtonCableMaterialCfg(
                     stretch_stiffness=1e3,
                     bend_stiffness=1e-4,
@@ -104,19 +113,45 @@ def design_scene(num_cables: int) -> dict[str, "CableObject | RigidObject"]:
                     bend_damping=1e-4,
                     density=100.0,
                 ),
-                collision_props=sim_utils.CollisionPropertiesCfg(),
             ),
-            init_state=CableObjectCfg.InitialStateCfg(pos=(cx, cy, cz), rot=z_axis_quat(angle)),
+            init_state=CableObjectCfg.InitialStateCfg(
+                pos=(0.0, 0.0, 0.5),
+            ),
             attachments=[
                 CableAttachmentCfg(
                     target_prim_path=plug_prim_path,
-                    cable_anchor="tail",
-                    local_pos=(0.0, 0.0, 0.0),
-                    local_quat=(1.0, 0.0, 0.0, 0.0),
+                    cable_anchor="head",
                 ),
             ],
         )
-        entities[f"Cable{idx:03d}"] = CableObject(cfg=cable_cfg)
+        entities[f"Cable1{idx:03d}"] = CableObject(cfg=cable_cfg)
+
+        # cable_cfg = CableObjectCfg(
+        #     prim_path=f"/World/Origin/Cable{idx:03d}",
+        #     spawn=sim_utils.CableCfg(
+        #         positions=[(i * segment_length, 0.0, 0.0) for i in range(num_points)],
+        #         width=width,
+        #         visual_material=sim_utils.PreviewSurfaceCfg(
+        #             diffuse_color=(random.random(), random.random(), random.random())
+        #         ),
+        #         physics_material=NewtonCableMaterialCfg(
+        #             stretch_stiffness=1e3,
+        #             bend_stiffness=1e-4,
+        #             stretch_damping=1e-1,
+        #             bend_damping=1e-4,
+        #             density=100.0,
+        #         ),
+        #         collision_props=sim_utils.CollisionPropertiesCfg(),
+        #     ),
+        #     init_state=CableObjectCfg.InitialStateCfg(pos=(cx, cy, cz)),
+        #     attachments=[
+        #         CableAttachmentCfg(
+        #             target_prim_path=plug_prim_path,
+        #             cable_anchor="tail",
+        #         ),
+        #     ],
+        # )
+        # entities[f"Cable{idx:03d}"] = CableObject(cfg=cable_cfg)
 
     return entities
 
@@ -165,10 +200,7 @@ def main():
         dt=0.01,
         device=args_cli.device,
         physics=physics_cfg,
-        visualizer_cfgs=[
-            NewtonVisualizerCfg(eye=(2.0, 2.0, 1.0), lookat=(0.0, 0.0, 0.25)),
-            KitVisualizerCfg(eye=(2.0, 2.0, 1.0), lookat=(0.0, 0.0, 0.25)),
-        ],
+        visualizer_cfgs=[NewtonVisualizerCfg(eye=(0.5, 1.5, 0.5), lookat=(0.0, 0.0, 0.05))],
     )
     sim = sim_utils.SimulationContext(sim_cfg)
