+++ source/isaaclab_contrib/isaaclab_contrib/cable/__init__.pyi
@@ -4,10 +4,11 @@
 # SPDX-License-Identifier: BSD-3-Clause
 
 __all__ = [
+    "CableAttachmentCfg",
     "CableObject",
     "CableObjectCfg",
     "CableRegistryEntry",
 ]
 
 from .cable_object import CableObject, CableRegistryEntry
-from .cable_object_cfg import CableObjectCfg
+from .cable_object_cfg import CableAttachmentCfg, CableObjectCfg
